fix: Changed the way the contact is created.
** Removed empty data.**

the previous way the client was creating it was not exactly suitable, so I decided to remove the idea of ​​creating an empty event to fill the database.

** Combination of 2 types of information **

The "events" and "notes" information was separate, this would be a problem when I established the order in which they were displayed. Taking into account that both cards are very similar, I decided to combine them into one to display directly from the most recent one.